Split trace and metrics marshalers to packages instead of huge classes.
