+++ exporters/otlp/common/src/main/java/io/opentelemetry/exporter/otlp/internal/MarshalerUtil.java
@@ -21,13 +21,19 @@ import java.util.Map;
 import java.util.function.Function;
 import javax.annotation.Nullable;
 
-final class MarshalerUtil {
+/**
+ * Marshaler utilities.
+ *
+ * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
+ * at any time.
+ */
+public final class MarshalerUtil {
   private static final int TRACE_ID_VALUE_SIZE =
       CodedOutputStream.computeLengthDelimitedFieldSize(TraceId.getLength() / 2);
   private static final int SPAN_ID_VALUE_SIZE =
       CodedOutputStream.computeLengthDelimitedFieldSize(SpanId.getLength() / 2);
 
-  static final boolean JSON_AVAILABLE;
+  private static final boolean JSON_AVAILABLE;
 
   static {
     boolean jsonAvailable = false;
@@ -42,11 +48,13 @@ final class MarshalerUtil {
 
   static final byte[] EMPTY_BYTES = new byte[0];
 
-  static <T, U> Map<Resource, Map<InstrumentationLibraryInfo, List<U>>> groupByResourceAndLibrary(
-      Collection<T> dataList,
-      Function<T, Resource> getResource,
-      Function<T, InstrumentationLibraryInfo> getInstrumentationLibrary,
-      Function<T, U> createMarshaler) {
+  /** Groups SDK items by resource and instrumentation library. */
+  public static <T, U>
+      Map<Resource, Map<InstrumentationLibraryInfo, List<U>>> groupByResourceAndLibrary(
+          Collection<T> dataList,
+          Function<T, Resource> getResource,
+          Function<T, InstrumentationLibraryInfo> getInstrumentationLibrary,
+          Function<T, U> createMarshaler) {
     // expectedMaxSize of 8 means initial map capacity of 16 to match HashMap
     IdentityHashMap<Resource, Map<InstrumentationLibraryInfo, List<U>>> result =
         new IdentityHashMap<>(8);
@@ -83,7 +91,8 @@ final class MarshalerUtil {
     return new String(jsonBytes, 1, jsonBytes.length - 2, StandardCharsets.UTF_8);
   }
 
-  static int sizeRepeatedFixed64(ProtoFieldInfo field, List<Long> values) {
+  /** Returns the size of a repeated fixed64 field. */
+  public static int sizeRepeatedFixed64(ProtoFieldInfo field, List<Long> values) {
     return sizeRepeatedFixed64(field, values.size());
   }
 
@@ -98,12 +107,15 @@ final class MarshalerUtil {
     return size;
   }
 
-  static int sizeRepeatedDouble(ProtoFieldInfo field, List<Double> values) {
+  /** Returns the size of a repeated double field. */
+  public static int sizeRepeatedDouble(ProtoFieldInfo field, List<Double> values) {
     // Same as fixed64.
     return sizeRepeatedFixed64(field, values.size());
   }
 
-  static <T extends Marshaler> int sizeRepeatedMessage(ProtoFieldInfo field, T[] repeatedMessage) {
+  /** Returns the size of a repeated message field. */
+  public static <T extends Marshaler> int sizeRepeatedMessage(
+      ProtoFieldInfo field, T[] repeatedMessage) {
     int size = 0;
     int fieldTagSize = field.getTagSize();
     for (Marshaler message : repeatedMessage) {
@@ -113,7 +125,9 @@ final class MarshalerUtil {
     return size;
   }
 
-  static int sizeRepeatedMessage(ProtoFieldInfo field, List<? extends Marshaler> repeatedMessage) {
+  /** Returns the size of a repeated message field. */
+  public static int sizeRepeatedMessage(
+      ProtoFieldInfo field, List<? extends Marshaler> repeatedMessage) {
     int size = 0;
     int fieldTagSize = field.getTagSize();
     for (Marshaler message : repeatedMessage) {
@@ -123,69 +137,79 @@ final class MarshalerUtil {
     return size;
   }
 
-  static int sizeMessage(ProtoFieldInfo field, Marshaler message) {
+  /** Returns the size of a message field. */
+  public static int sizeMessage(ProtoFieldInfo field, Marshaler message) {
     int fieldSize = message.getBinarySerializedSize();
     return field.getTagSize() + CodedOutputStream.computeUInt32SizeNoTag(fieldSize) + fieldSize;
   }
 
-  static int sizeBool(ProtoFieldInfo field, boolean value) {
+  /** Returns the size of a bool field. */
+  public static int sizeBool(ProtoFieldInfo field, boolean value) {
     if (!value) {
       return 0;
     }
     return field.getTagSize() + CodedOutputStream.computeBoolSizeNoTag(value);
   }
 
-  static int sizeUInt32(ProtoFieldInfo field, int message) {
+  /** Returns the size of a uint32 field. */
+  public static int sizeUInt32(ProtoFieldInfo field, int message) {
     if (message == 0) {
       return 0;
     }
     return field.getTagSize() + CodedOutputStream.computeUInt32SizeNoTag(message);
   }
 
-  static int sizeDouble(ProtoFieldInfo field, double value) {
+  /** Returns the size of a double field. */
+  public static int sizeDouble(ProtoFieldInfo field, double value) {
     if (value == 0D) {
       return 0;
     }
     return field.getTagSize() + CodedOutputStream.computeDoubleSizeNoTag(value);
   }
 
-  static int sizeFixed64(ProtoFieldInfo field, long message) {
+  /** Returns the size of a fixed64 field. */
+  public static int sizeFixed64(ProtoFieldInfo field, long message) {
     if (message == 0L) {
       return 0;
     }
     return field.getTagSize() + CodedOutputStream.computeFixed64SizeNoTag(message);
   }
 
-  static int sizeBytes(ProtoFieldInfo field, byte[] message) {
+  /** Returns the size of a bytes field. */
+  public static int sizeBytes(ProtoFieldInfo field, byte[] message) {
     if (message.length == 0) {
       return 0;
     }
     return field.getTagSize() + CodedOutputStream.computeByteArraySizeNoTag(message);
   }
 
+  /** Returns the size of a enum field. */
   // Assumes OTLP always defines the first item in an enum with number 0, which it does and will.
-  static int sizeEnum(ProtoFieldInfo field, int value) {
+  public static int sizeEnum(ProtoFieldInfo field, int value) {
     if (value == 0) {
       return 0;
     }
     return field.getTagSize() + CodedOutputStream.computeEnumSizeNoTag(value);
   }
 
-  static int sizeTraceId(ProtoFieldInfo field, @Nullable String traceId) {
+  /** Returns the size of a trace_id field. */
+  public static int sizeTraceId(ProtoFieldInfo field, @Nullable String traceId) {
     if (traceId == null) {
       return 0;
     }
     return field.getTagSize() + TRACE_ID_VALUE_SIZE;
   }
 
-  static int sizeSpanId(ProtoFieldInfo field, @Nullable String spanId) {
+  /** Returns the size of a span_id field. */
+  public static int sizeSpanId(ProtoFieldInfo field, @Nullable String spanId) {
     if (spanId == null) {
       return 0;
     }
     return field.getTagSize() + SPAN_ID_VALUE_SIZE;
   }
 
-  static byte[] toBytes(@Nullable String value) {
+  /** Converts the string to utf8 bytes for encoding. */
+  public static byte[] toBytes(@Nullable String value) {
     if (value == null || value.isEmpty()) {
       return EMPTY_BYTES;
     }
