+++ exporters/otlp/metrics/src/main/java/io/opentelemetry/exporter/otlp/metrics/OtlpGrpcMetricExporter.java
@@ -11,8 +11,8 @@ import com.google.common.util.concurrent.MoreExecutors;
 import io.grpc.Codec;
 import io.grpc.ManagedChannel;
 import io.grpc.Status;
-import io.opentelemetry.exporter.otlp.internal.MetricsRequestMarshaler;
 import io.opentelemetry.exporter.otlp.internal.grpc.ManagedChannelUtil;
+import io.opentelemetry.exporter.otlp.internal.metrics.MetricsRequestMarshaler;
 import io.opentelemetry.sdk.common.CompletableResultCode;
 import io.opentelemetry.sdk.internal.ThrottlingLogger;
 import io.opentelemetry.sdk.metrics.data.MetricData;
