+++ exporters/otlp/trace/src/main/java/io/opentelemetry/exporter/otlp/trace/MarshalerTraceServiceGrpc.java
@@ -8,8 +8,8 @@ package io.opentelemetry.exporter.otlp.trace;
 import static io.grpc.MethodDescriptor.generateFullMethodName;
 
 import io.grpc.MethodDescriptor;
-import io.opentelemetry.exporter.otlp.internal.TraceRequestMarshaler;
 import io.opentelemetry.exporter.otlp.internal.grpc.MarshalerInputStream;
+import io.opentelemetry.exporter.otlp.internal.traces.TraceRequestMarshaler;
 import java.io.InputStream;
 
 // Adapted from the protoc generated code for TraceServiceGrpc.
